requirements.txt
aioice==0.9.0
aiortc==1.5.0
alabaster==0.7.13
altair==4.2.2
asttokens==2.2.1
attrs==23.1.0
av==10.0.0
Babel==2.12.1
backcall==0.2.0
blinker==1.6.2
botocore==1.31.39
branca==0.6.0
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.2
colorama==0.3.9
comm==0.1.4
contourpy==1.1.0
coverage==7.3.0
cryptography==41.0.3
cycler==0.10.0
debugpy==1.6.7.post1
decorator==5.1.1
dnspython==2.4.2
docutils==0.20.1
entrypoints==0.4
executing==1.2.0
filelock==3.12.3
flake8==6.1.0
folium==0.14.0
fonttools==4.42.1
gitdb==4.0.10
GitPython==3.1.33
gcloud==0.18.3
google-api-python-client==1.6
google-cloud-storage==2.10.0
google-crc32c==1.5.0
httplib2==0.15.0
idna==2.10
ifaddr==0.2.0
imagesize==1.4.1
importlib-metadata==6.8.0
ipykernel==6.25.1
ipython==8.14.0
jedi==0.19.0
Jinja2==3.1.2
jmespath==0.10.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client==8.3.1
jupyter_core==5.3.1
kiwisolver==1.4.5
lit==16.0.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
nest-asyncio==1.5.7
networkx==3.1
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
packaging==23.1
pandas==2.1.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==13.0.0
pyasn1==0.5.0
pycodestyle==2.11.0
pycparser==2.21
pydeck==0.8.1b0
pyee==11.0.0
pyflakes==3.1.0
Pygments==2.16.1
pylibsrtp==0.8.0
Pympler==1.0.1
pyOpenSSL==23.2.0
pyasn1==0.5.0
pycodestyle==2.11.0
pyflakes==3.1.0
Pygments==2.16.1
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.5.2
roboflow==1.1.4
rpds-py==0.10.0
requests==2.31.0
requests-toolbelt==1.0.0
roboflow==1.1.4
rsa==3.4.2
s3transfer==0.2.1
scipy==1.11.2
seaborn==0.12.2
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
Sphinx==7.2.5
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
stack-data==0.6.2
streamlit==1.22.0
streamlit-camera-input-live==0.2.0
streamlit-folium==0.13.0
streamlit-webrtc==0.47.0
supervision==0.14.0
sympy==1.12
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
supervision==0.14.0
sympy==1.12
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
ultralytics==8.0.167
urllib3==1.26.16
validators==0.21.2
watchdog==3.0.0
wcwidth==0.2.6
wget==3.2
zipp==3.16.2
ultralytics==8.0.167
wcwidth==0.2.6
wget==3.2